Reformat all te things
