+++ Sources/KonaneGame.swift
@@ -33,48 +33,3 @@ class KonaneGame {
     
     
 }
-
-func initializeGame() -> (blackIsHuman: Bool, whiteIsHuman: Bool) {
-    var isBlackHuman: Bool = true
-    var isWhiteHuman: Bool = true
-    var response = ""
-    print("Is black human? [Y/N]:")
-    
-    response = String(readLine()!)!
-    if response ==  "Y" {
-        print("Black is human")
-        isBlackHuman = true
-    } else if response ==  "N" {
-        print("Black is AI")
-        isBlackHuman = false
-    } else {
-        print("Invalid. Please start over.")
-    }
-    
-    
-    print("Is white human? [Y/N]")
-    
-    response = String(readLine()!)!
-    if response ==  "Y" {
-        print("White is human")
-        isWhiteHuman = true
-    } else if response ==  "N" {
-        print("White is AI")
-        isWhiteHuman = false
-    } else {
-        print("Invalid. Please start over.")
-        
-    }
-    
-    print()
-    
-    return (isBlackHuman, isWhiteHuman)
-    //END User input to initialize KonaneGame
-    
-    //Initialize KonaneGame with user input
-    //WILL FIX ONCE FUNC IS DONE
-    //init (blackIsHuman = initializeBlack, whiteIsHuman = initializeWhite)
-    
-
-}
-
