+++ Sources/KonaneMove.swift
@@ -32,26 +32,3 @@ class KonaneMove {
 	}
 }
 
-var myKonaneGameState = KonaneGameState.konaneGameState
-
-//Actually programming a game move
-//NOT WORKING YET ERRORS
-func searchJumpPiece() {
-    
-    var searchTable: [[Int]] = []
-
-    for column in myKonaneGameState.gameBoard {
-        for row in myKonaneGameState.gameBoard[column] {
-            if myKonaneGameState.isBlackTurn {
-                if myKonaneGameState.gameBoard[column][row] == konaneColor.black {
-                    searchTable.append([column, row])
-                }
-            }
-            else if !myKonaneGameState.isBlackTurn {
-                if myKonaneGameState.gameBoard[[column][row] == konaneColor.white {
-                    searchTable.append([column, row])
-                }
-            }
-        }
-    }
-}
